Bump version to 1.0.0.1
This is just so that the plugin can be PRed and it will ensure the manifest and plugin package is rebuilt.

I've disabled the copy to devPlugins operation as we may be phasing that out in the future and it should not be relied upon. For testing purposes, use the external devplugins folder feature in Dalamud Settings instead.